camera-recorder 05.10/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python==4.9.0.80
requests==2.31.0
PyYAML==6.0.1
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
